requirements.txt
google-generativeai
pandas
openpyxl
python-dotenv

google-cloud-aiplatform>=1.55.0
